Moving category and categories components to pages and plug the routes of the header
