+++ src/app/header/header.component.html
@@ -7,12 +7,12 @@
         <nav>
             <ul class="divnav">
                 <li><a [routerLink]="APP_ROUTES.index">Accueil</a></li>
-                <li><a class="categ" [routerLink]="">Catégories</a></li>
-                <li><a id="fun" [routerLink]="">Hav'Fun</a></li>
-                <li><a id="sport" [routerLink]="">NRgy</a></li>
-                <li><a id="carie" [routerLink]="">Carie</a></li>
-                <li><a id="miam" [routerLink]="">Miam</a></li>
-                <li><a id="money" [routerLink]="">Ca coute bonbon</a></li>
+                <li><a class="categ" [routerLink]="APP_ROUTES.categories">Catégories</a></li>
+                <li><a id="fun" [routerLink]="APP_ROUTES.categories+'/1'">Hav'Fun</a></li>
+                <li><a id="sport" [routerLink]="APP_ROUTES.categories+'/2'">NRgy</a></li>
+                <li><a id="carie" [routerLink]="APP_ROUTES.categories+'/3'">Carie</a></li>
+                <li><a id="miam" [routerLink]="APP_ROUTES.categories+'/4'">Miam</a></li>
+                <li><a id="money" [routerLink]="APP_ROUTES.categories+'/5'">Ca coute bonbon</a></li>
                 <li><a id="login" [routerLink]="">Connexion</a></li>
                 <li></li>
             </ul>
