+++ src/app/pages/categories/categories.component.html
