+++ src/app/pages/categories/categories.component.scss
