+++ src/app/pages/categories/categories.component.spec.ts
