+++ src/app/pages/categories/categories.component.ts
